+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support to delete files in S3 / object storage through the `php bin/console database:drop` command, closes #309.
 - Add support for `Content-Disposition` HTTP header, including transliteration of non-latin-characters, closes #447.
 - Add dependency to PHP library `cardinalby/content-disposition`.
+- Add 'reserved type' exception; nodes of type 'User', 'Token' and 'Upload' can not be manually created through generic
+  API endpoints.
 
 ### Changed
 - Upgrade FrankenPHP to 1.12.1, related to #457 and #461.